[CI] Skip pairing if possible when using chip-tool or darwin-framework-tool to make CI faster
